+++ mzs_tools/plugin_utils/settings.py
@@ -18,6 +18,7 @@ class PlgSettingsStructure:
     # global
     debug_mode: bool = False
     auto_advanced_editing: bool = True
+    java_home_path: str = ""
     version: str = __version__
 
 
@@ -57,8 +58,9 @@ class PlgOptionsManager:
         :return: plugin settings value matching key
         """
         if not hasattr(PlgSettingsStructure, key):
+            field_names = [field.name for field in fields(PlgSettingsStructure)]
             logging.MzSToolsLogger.log(
-                message="Bad settings key. Must be one of: {}".format(",".join(PlgSettingsStructure._fields)),
+                message="Bad settings key. Must be one of: {}".format(",".join(field_names)),
                 log_level=1,
             )
             return None
@@ -88,8 +90,9 @@ class PlgOptionsManager:
         :rtype: bool
         """
         if not hasattr(PlgSettingsStructure, key):
+            field_names = [field.name for field in fields(PlgSettingsStructure)]
             logging.MzSToolsLogger.log(
-                message="Bad settings key. Must be one of: {}".format(",".join(PlgSettingsStructure._fields)),
+                message="Bad settings key. Must be one of: {}".format(",".join(field_names)),
                 log_level=2,
             )
             return False