+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## [2.0.5-beta3] - 2025-11-08
 
 - Fix for HVSR view to handle comma as decimal separator
-- Custom python dependency manager and new tool to check and install required python libraries (alternative to QPIP)
+- Custom dependency manager and new tool to check and install required python libraries (alternative to QPIP) and detect Java JVM
+- New setting to manually store the Java JRE location
 - Disabled plugin dependency on QPIP
 - Implemented a db_manager module to improve database operations
 - Handle malformed date strings for data export to Access
-- Export tool improvements
+- Data import and export improvements
 
 ## [2.0.4] - 2025-09-18